+++ tests/HelperBracketTrait/HelperBracketDeleteTest.php
@@ -62,6 +62,9 @@ final class HelperBracketDeleteTest extends TestCase
         $string = Helper::bracketDelete($example3, '<a', '>', 2);
         $this->assertEquals($string, '<a href="1"><a href="2"></a></a> </a>');
 
+        $string = Helper::bracketDelete('abc', '(', ')', 0);
+        $this->assertEquals($string, 'abc');
+
         $string = Helper::bracketDelete(null, '(', ')', 0);
         $this->assertEquals($string, '');
     }