+++ tests/HelperNumberTrait/HelperNumberFromStringTest.php
@@ -17,12 +17,17 @@ final class HelperNumberFromStringTest extends TestCase
     #[Test]
     public function numberFromString(): void
     {
-        // $integer = Helper::numberFromString('сто двадцать три');
-        // $this->assertTrue($integer === 123);
+        $integer = Helper::numberFromString('сто двадцать три');
+        $this->assertSame($integer, 123);
 
         $integer = Helper::numberFromString('сто двадцать три целых четыреста пятьдесят шесть тысячных');
-        $this->assertEquals($integer, 123.456);
-        $this->assertTrue($integer === 123.456);
+        $this->assertSame($integer, 123.456);
+
+        $integer = Helper::numberFromString('пять целых одна миллионная');
+        $this->assertSame($integer, 5.000001);
+
+        $integer = Helper::numberFromString('минус миллион один');
+        $this->assertSame($integer, -1000001);
 
         $integer = Helper::numberFromString(null);
         $this->assertEquals($integer, 0);