+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 ## Версия 1.13
 - Добавлен метод Helper::bracketChange
 - Добавлен метод Helper::bracketCopy
+- Добавлен метод Helper::bracketCount
 - Добавлен метод Helper::bracketDelete
 - Добавлен метод Helper::bracketReplace
 - Добавлен метод Helper::numberFromString
+- Добавлен метод Helper::numberSwap
+
+- Обновлен метод Helper::stringDeleteMultiples
 
 ## Версия 1.12
 - Добавлен метод Helper::cacheRuntimeClear